Copy over static resource handler to new WebDriverServlet.
And set the headers for CORS too.
