+++ java/server/src/org/openqa/selenium/remote/server/DriverServlet.java
@@ -18,12 +18,8 @@
 package org.openqa.selenium.remote.server;
 
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Strings.nullToEmpty;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.io.ByteStreams;
-import com.google.common.io.Files;
 import com.google.common.net.HttpHeaders;
 import com.google.common.net.MediaType;
 
@@ -36,9 +32,6 @@ import org.openqa.selenium.remote.server.xdrpc.CrossDomainRpcLoader;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.URL;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import java.util.logging.Handler;
@@ -244,61 +237,4 @@ public class DriverServlet extends HttpServlet {
       return (DriverSessions) attribute;
     }
   }
-
-  private static class StaticResourceHandler {
-    private static final ImmutableMap<String, MediaType> MIME_TYPES = ImmutableMap.of(
-        "css", MediaType.CSS_UTF_8.withoutParameters(),
-        "html", MediaType.HTML_UTF_8.withoutParameters(),
-        "js", MediaType.JAVASCRIPT_UTF_8.withoutParameters());
-
-    private static final String STATIC_RESOURCE_BASE_PATH = "/static/resource/";
-    private static final String HUB_HTML_PATH = STATIC_RESOURCE_BASE_PATH + "hub.html";
-
-    public boolean isStaticResourceRequest(HttpServletRequest request) {
-      return "GET".equalsIgnoreCase(request.getMethod())
-             && nullToEmpty(request.getPathInfo()).startsWith(STATIC_RESOURCE_BASE_PATH);
-    }
-
-    public void redirectToHub(HttpServletRequest request, HttpServletResponse response)
-        throws IOException {
-      response.sendRedirect(request.getContextPath() + request.getServletPath() + HUB_HTML_PATH);
-    }
-
-    public void service(HttpServletRequest request, HttpServletResponse response)
-        throws IOException {
-      checkArgument(isStaticResourceRequest(request));
-
-      String path = String.format(
-          "/%s/%s",
-          StaticResourceHandler.class.getPackage().getName().replace(".", "/"),
-          request.getPathInfo().substring(STATIC_RESOURCE_BASE_PATH.length()));
-      URL url = StaticResourceHandler.class.getResource(path);
-
-      if (url == null) {
-        response.sendError(HttpServletResponse.SC_NOT_FOUND);
-        return;
-      }
-
-      response.setStatus(HttpServletResponse.SC_OK);
-
-      String extension = Files.getFileExtension(path);
-      if (MIME_TYPES.containsKey(extension)) {
-        response.setContentType(MIME_TYPES.get(extension).toString());
-      }
-
-      byte[] data = getResourceData(url);
-      response.setContentLength(data.length);
-
-      try (OutputStream output = response.getOutputStream()) {
-        output.write(data);
-      }
-    }
-
-    private byte[] getResourceData(URL url) throws IOException {
-      try (InputStream stream = url.openStream()) {
-        return ByteStreams.toByteArray(stream);
-      }
-    }
-  }
-
 }
